requirements.txt
azure-cognitiveservices-vision-customvision
msrest
matplotlib
pillow
numpy

